pyproject.toml
[project]
name = "oncp-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
